% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_1213.R
\name{load_sitreps_1213}
\alias{load_sitreps_1213}
\title{Load timeseries of winter situation reports from 2012-13}
\usage{
load_sitreps_1213(
  sitrep_url =
    "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2013/04/Daily-SR-Timeseries-.xls"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
Downloads and cleans the "Daily SR – Timeseries 6 November 2012 – 28 February 2013.xls" data from https://www.england.nhs.uk/statistics/statistical-work-areas/winter-daily-sitreps/winter-sitrep-data-2012-13/
}
\details{
The returned tibble contains:
- A&E diverts
- A&E closures
- Bed occupancy
- Beds closed due to diarrhoea, vomiting, norovirus
}
