% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_1314.R
\name{load_sitreps_1314}
\alias{load_sitreps_1314}
\title{Load timeseries of winter situation reports from 2013-14}
\usage{
load_sitreps_1314(
 
    sitrep_url = "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2013/11/DailySR-Timeseries-WE-30.03.14.xlsx"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
Downloads and cleans the "Daily SR – Time series 4 November 2013 to 30 March 2014" data from https://www.england.nhs.uk/statistics/statistical-work-areas/winter-daily-sitreps/winter-daily-sitrep-2013-14-data-2/
}
\details{
The returned tibble contains:
- A&E diverts
- A&E closures
- Bed occupancy
- Beds closed due to diarrhoea, vomiting, norovirus
}
