% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_1415.R
\name{load_sitreps_1415}
\alias{load_sitreps_1415}
\title{Load timeseries of winter situation reports from 2014-15}
\usage{
load_sitreps_1415(
  sitrep_url =
    "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2014/12/DailySR-Timeseries-WE-29.03.15.xlsx"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
Downloads and cleans the "Winter SitRep Part A: Acute Time Series 03 November 2014 to 29 March 2015" data from https://www.england.nhs.uk/statistics/statistical-work-areas/winter-daily-sitreps/winter-daily-sitrep-2014-15-data/
}
\details{
The returned tibble contains:
- A&E diverts
- A&E closures
- Bed occupancy
- Beds closed due to diarrhoea, vomiting, norovirus
}
