% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_1516.R
\name{load_sitreps_1516}
\alias{load_sitreps_1516}
\title{Load timeseries of winter situation reports from 2015-16}
\usage{
load_sitreps_1516(
 
    sitrep_url = "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2015/12/DailySR-Timeseries-WE-28.02.16.xlsx"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
Downloads and cleans the "Winter SitRep Part A: Acute Time Series 30 November 2015 to 28 February 2016" data from https://www.england.nhs.uk/statistics/statistical-work-areas/winter-daily-sitreps/winter-daily-sitrep-2015-16-data/
}
\details{
The returned tibble contains:
- A&E diverts
- A&E closures
- A&E attendance
- Bed occupancy
- Beds closed due to diarrhoea, vomiting, norovirus
}
