% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_1617.R
\name{load_sitreps_1617}
\alias{load_sitreps_1617}
\title{Load timeseries of winter situation reports from 2016-17}
\usage{
load_sitreps_1617(
  sitrep_url =
    "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2016/12/DailySR-Web-file-Time-Series-18.xlsx"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
Downloads and cleans the "Winter SitRep Part A: Acute Time Series 1 December 2016 to 12 March 2017" data from https://www.england.nhs.uk/statistics/statistical-work-areas/winter-daily-sitreps/winter-daily-sitrep-2016-17-data/
}
\details{
The returned tibble contains:
- A&E diverts
- A&E closures
- Bed occupancy
- Beds closed due to diarrhoea, vomiting, norovirus
}
