% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_1718.R
\name{load_sitreps_1819}
\alias{load_sitreps_1819}
\title{Load timeseries of winter situation reports from 2018-19}
\usage{
load_sitreps_1819(
 
    sitrep_url = "https://www.england.nhs.uk/statistics/wp-content/uploads/sites/2/2019/03/Winter-data-timeseries-20190307.xlsx"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
Downloads and cleans the "Winter SitRep: Acute Time series 3 December 2018 to 3 March 2019" data from https://www.england.nhs.uk/statistics/statistical-work-areas/winter-daily-sitreps/winter-daily-sitrep-2018-19-data/
}
\details{
The returned tibble contains:
- A&E diverts
- A&E closures
- Ambulance delays
- Bed occupancy
- Beds occupied by long-stay patients (> 7 days and > 21 days)
- Beds closed due to diarrhoea, vomiting, norovirus
}
