% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_2021.R
\name{load_sitreps_2021}
\alias{load_sitreps_2021}
\title{Load timeseries of winter situation reports for 2020-21}
\usage{
load_sitreps_2021(sitrep_url)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
The returned tibble contains:
- General & Acute beds
- Adult critical care
}
