% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_2122.R
\name{load_sitreps_2122}
\alias{load_sitreps_2122}
\title{Load timeseries of winter situation reports for 2021-22}
\usage{
load_sitreps_2122(sitrep_url)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}
}
\description{
The returned tibble contains:
- General & Acute beds
- Adult critical care
}
