% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_2223.R
\name{load_sitreps_2223}
\alias{load_sitreps_2223}
\title{Load timeseries of winter situation reports for 2022-23}
\usage{
load_sitreps_2223(sitrep_url, ambo_url)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}

\item{ambo_url}{URL for ambulance sitrep timeseries}
}
\description{
The returned tibble contains:
- General & Acute beds
- Adult critical care
}
