% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sitreps_generic.R
\name{load_sitreps_generic}
\alias{load_sitreps_generic}
\title{Load timeseries of winter situation reports in its most up-to-date format (from 2017-18 to present)}
\usage{
load_sitreps_generic(
  sitrep_url,
  closures_sheet_name = "A&E Closures",
  diverts_sheet_name = "A&E Diverts"
)
}
\arguments{
\item{sitrep_url}{URL of the timeseries file}

\item{closures_sheet_name}{Name of the 'A&E Closures' worksheet (it can differ from winter to winter)}

\item{diverts_sheet_name}{Name of the 'A&E Diverts' worksheet (it can differ from winter to winter)}
}
\description{
The returned tibble contains:
- A&E diverts
- A&E closures
- Ambulance delays
- Bed occupancy
- Beds occupied by long-stay patients (> 7 days and > 21 days)
- Beds closed due to diarrhoea, vomiting, norovirus
}
