% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_trusts.R
\name{load_trusts}
\alias{load_trusts}
\title{Load list of NHS Trust locations}
\usage{
load_trusts(url = "https://nhsenglandfilestore.s3.amazonaws.com/ods/etr.csv")
}
\arguments{
\item{url}{Location of the NHS Trusts and Sites .csv file}
}
\description{
Uses NHS England's data on Trusts and Sites from https://data.england.nhs.uk/dataset/ods-nhs-trusts-and-sites
}
