% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep1516}
\alias{sitrep1516}
\title{Winter sitrep for 2015-16}
\format{
A data frame with 9,486 rows and 14 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep1516
}
\description{
Winter sitrep for 2015-16
}
\keyword{datasets}
