% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep1617}
\alias{sitrep1617}
\title{Winter sitrep for 2016-17}
\format{
A data frame with 15,606 rows and 13 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep1617
}
\description{
Winter sitrep for 2016-17
}
\keyword{datasets}
