% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep1718}
\alias{sitrep1718}
\title{Winter sitrep for 2017-18}
\format{
A data frame with 14,490 rows and 17 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep1718
}
\description{
Winter sitrep for 2017-18
}
\keyword{datasets}
