% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep1819}
\alias{sitrep1819}
\title{Winter sitrep for 2018-19}
\format{
A data frame with 12,285 rows and 17 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep1819
}
\description{
Winter sitrep for 2018-19
}
\keyword{datasets}
