% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep1920}
\alias{sitrep1920}
\title{Winter sitrep for 2019-20}
\format{
A data frame with 12,103 rows and 17 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep1920
}
\description{
Winter sitrep for 2019-20
}
\keyword{datasets}
