% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep2021}
\alias{sitrep2021}
\title{Winter sitrep for 2020-21}
\format{
A data frame with 23,023 rows and 15 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep2021
}
\description{
Winter sitrep for 2020-21
}
\keyword{datasets}
