% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep2122}
\alias{sitrep2122}
\title{Winter sitrep for 2021-22}
\format{
A data frame with 18,620 rows and 17 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep2122
}
\description{
Winter sitrep for 2021-22
}
\keyword{datasets}
