% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitrep2223}
\alias{sitrep2223}
\title{Winter sitrep for 2022-23}
\format{
A data frame with 6,580 rows and 19 variables
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
sitrep2223
}
\description{
Winter sitrep for 2022-23
}
\keyword{datasets}
